pyproject.toml
[project]
name = "authorization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.17.1",
    "fastapi>=0.119.0",
    "openai>=1.45.0",
    "pipreqs>=0.5.0",
    "uvicorn>=0.37.0",
]
